requirements.txt
click==8.1.7
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
questionary==2.0.1
newspaper4k
lxml_html_clean==0.2.0
spacy==3.7.2
tabulate==0.9.0
youtube-transcript-api==0.6.2
urllib3==2.1.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
tabulate
supabase
